Convert an Excel brickr template into a brickr 3D object

Convert an Excel brickr template into a brickr 3D object

bricks_from_excel(excel_table, repeat_levels = 1, increment_level = 0,
  max_level = Inf, increment_x = 0, max_x = Inf, increment_y = 0,
  max_y = Inf, exclude_color = NULL, exclude_level = NULL)

Arguments

excel_table
Sheet imported from a brickr Excel template to build model. Contains stud placement and colors.

repeat_levels
How many times to repeat a level. Can save time in model planning. Default is 1.

increment_level
Default '0'. Use in animations. Shift Level/z dimension by an integer.

max_level
Default 'Inf'. Use in animations. Any Level/z values above this value will be cut off.

increment_x
Default '0'. Use in animations. Shift x dimension by an integer.

max_x
Default 'Inf'. Use in animations. Any x values above this value will be cut off.

increment_y
Default '0'. Use in animations. Shift y dimension by an integer.

max_y
Default 'Inf'. Use in animations. Any y values above this value will be cut off.

exclude_color
Numeric array of color ID numbers to exclude.

exclude_level
Numeric array of Level/z dimensions to exclude.

Value

A list with elements Img_lego to pass to collect_bricks().

See also