Build a brickr 3D object from rayshader hillshade & heightmap matrices

Build a brickr 3D object from rayshader hillshade & heightmap matrices

bricks_from_rayshader(hillshade, heightmap, max_height = 12,
  img_size = 48)

Arguments

hillshade
Same as rayshader::plot_3d(). Hillshade/image to be added to 3D surface map.

heightmap
Same as rayshader::plot_3d(). A two-dimensional matrix, where each entry in the matrix is the elevation at that point. All points are assumed to be evenly spaced.

max_height
Maximum height of plot in LEGO bricks or plates.

img_size
Size of output image in pixel, where one pixel = one 'brick'. Use a single value (e.g. 48) for a square image with 48 pixels on each side. Use an array of two values for a rectangular image c(width, height).

Value

A list with elements threed_elevation and threed_hillshade to created 3D mosiacs with the rayshader package.

See also